Move common/ledger/util/Buffer to consumer
- Relocate a small helper class out of the "common" utility package and
  into the single consuming package.
- Stop exporting the Buffer type as it's a detail of the consumer.
