+++ common/ledger/blkstorage/fsblkstorage/protobuf_util.go
@@ -14,27 +14,27 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package fsblkstorage
 
 import (
 	"github.com/golang/protobuf/proto"
 	"github.com/pkg/errors"
 )
 
-// Buffer provides a wrapper on top of proto.Buffer.
+// buffer provides a wrapper on top of proto.Buffer.
 // The purpose of this wrapper is to get to know the current position in the []byte
-type Buffer struct {
+type buffer struct {
 	buf      *proto.Buffer
 	position int
 }
 
-// NewBuffer constructs a new instance of Buffer
-func NewBuffer(b []byte) *Buffer {
-	return &Buffer{proto.NewBuffer(b), 0}
+// newBuffer constructs a new instance of Buffer
+func newBuffer(b []byte) *buffer {
+	return &buffer{proto.NewBuffer(b), 0}
 }
 
 // DecodeVarint wraps the actual method and updates the position
-func (b *Buffer) DecodeVarint() (uint64, error) {
+func (b *buffer) DecodeVarint() (uint64, error) {
 	val, err := b.buf.DecodeVarint()
 	if err == nil {
 		b.position += proto.SizeVarint(val)
@@ -45,7 +45,7 @@ func (b *Buffer) DecodeVarint() (uint64, error) {
 }
 
 // DecodeRawBytes wraps the actual method and updates the position
-func (b *Buffer) DecodeRawBytes(alloc bool) ([]byte, error) {
+func (b *buffer) DecodeRawBytes(alloc bool) ([]byte, error) {
 	val, err := b.buf.DecodeRawBytes(alloc)
 	if err == nil {
 		b.position += proto.SizeVarint(uint64(len(val))) + len(val)
@@ -56,6 +56,6 @@ func (b *Buffer) DecodeRawBytes(alloc bool) ([]byte, error) {
 }
 
 // GetBytesConsumed returns the offset of the current position in the underlying []byte
-func (b *Buffer) GetBytesConsumed() int {
+func (b *buffer) GetBytesConsumed() int {
 	return b.position
 }
