+++ common/ledger/blkstorage/fsblkstorage/protobuf_util_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package fsblkstorage
 
 import (
 	"testing"
@@ -34,7 +34,7 @@ func TestBuffer(t *testing.T) {
 	pb.EncodeVarint(1000000)
 	pos4 := len(pb.Bytes())
 
-	b := NewBuffer(pb.Bytes())
+	b := newBuffer(pb.Bytes())
 	b.DecodeVarint()
 	assert.Equal(t, pos1, b.GetBytesConsumed())
 	b.DecodeRawBytes(false)
